Rename Sessionsocketservice
