+++ frontend/peerprep/package.json
@@ -52,6 +52,7 @@
     "react-markdown": "^9.0.1",
     "react-quill": "^2.0.0",
     "react-simple-code-editor": "^0.14.1",
+    "react-toastify": "^10.0.6",
     "remark-breaks": "^4.0.0",
     "remark-gfm": "^4.0.0",
     "socket.io-client": "^4.8.0",
